rename and move bucket guard to common
