+++ src/server/journal/streamer.cc
@@ -212,7 +212,7 @@ void RestoreStreamer::Start(util::FiberSocketBase* dest, bool send_lsn) {
 
     bool written = false;
     cursor = pt->Traverse(cursor, [&](PrimeTable::bucket_iterator it) {
-      BucketSerializationGuard guard(&bucket_ser_);
+      ConditionGuard guard(&bucket_ser_);
 
       db_slice_->FlushChangeToEarlierCallbacks(0 /*db_id always 0 for cluster*/,
                                                DbSlice::Iterator::FromPrime(it), snapshot_version_);
@@ -311,7 +311,7 @@ bool RestoreStreamer::WriteBucket(PrimeTable::bucket_iterator it) {
 void RestoreStreamer::OnDbChange(DbIndex db_index, const DbSlice::ChangeReq& req) {
   DCHECK_EQ(db_index, 0) << "Restore migration only allowed in cluster mode in db0";
 
-  BucketSerializationGuard guard(&bucket_ser_);
+  ConditionGuard guard(&bucket_ser_);
 
   PrimeTable* table = db_slice_->GetTables(0).first;
